rename: "service/user -> service/user.service.js"
